+++ R/form_retype.R
@@ -3,8 +3,8 @@
 #' Modify [type][meta_type()] of pdqr-function using method of choice.
 #'
 #' @param f A pdqr-function.
-#' @param type A desired type of output. Should be one of "discrete" or
-#'   "continuous".
+#' @param type A desired type of output. Should be one of "discrete",
+#'   "continuous", "value".
 #' @param method Retyping method. Should be one of "piecelin" or "dirac".
 #'
 #' @details If type of `f` is equal to input `type` then `f` is returned.
@@ -34,6 +34,20 @@
 #' probability taken from corresponding value of "prob" column. Output
 #' essentially represents a mixture of dirac-like distributions.
 #'
+#' Method "value" uses renormalized columns of `f`'s "x_tbl" metadata as values
+#' for output's "x_tbl" metadata. In other words, it preserves ratios between
+#' values of d-function at certain "x" points. Its main advantages are that this
+#' method can work well with any pdqr type and that two consecutive conversions
+#' return the same function. Conversion algorithm is as follows:
+#' - Retyping from "continuous" to `type` "discrete" is done by creating
+#' pdqr-function of corresponding class with the following "x_tbl" metadata: "x"
+#' column is the same as in `f`; "prob" column is equal to `f`'s "y" column
+#' after renormalization (so that their sum is 1).
+#' - Retyping from "discrete" to `type` "continuous" is done in the same
+#' fashion: "x" column is the same; "y" column is equal to `f`'s "prob" column
+#' after renormalization (so that total integral of piecewise-linear density is
+#' equal to 1).
+#'
 #' @return A pdqr-function with type equal to input `type`.
 #'
 #' @seealso [form_regrid()] for changing grid (rows of "x_tbl" metadata) of
@@ -59,13 +73,24 @@
 #' plot(my_con)
 #' lines(my_con_2, col = "blue")
 #'
+#' # Method "value"
+#' meta_x_tbl(form_retype(my_con, "discrete", method = "value"))
+#'
+#'   # This method returns input function after two consecutive conversions
+#' meta_x_tbl(
+#'   form_retype(
+#'     f = form_retype(my_dis, "continuous", method = "value"),
+#'     "discrete", method = "value"
+#'   )
+#' )
+#'
 #' @export
 form_retype <- function(f, type, method = "piecelin") {
   assert_pdqr_fun(f)
   assert_missing(type, "pdqr type of output")
   assert_pdqr_type(type)
   assert_type(method, is_string)
-  assert_in_set(method, c("piecelin", "dirac"))
+  assert_in_set(method, c("piecelin", "dirac", "value"))
 
   switch(
     type,
@@ -82,7 +107,8 @@ retype_dis <- function(f, method) {
   switch(
     method,
     piecelin = retype_dis_piecelin(f),
-    dirac = retype_dis_dirac(f)
+    dirac = retype_dis_dirac(f),
+    value = retype_dis_vlaue(f)
   )
 }
 
@@ -145,6 +171,14 @@ retype_dis_dirac <- function(f) {
   new_pdqr_by_ref(f)(data.frame(x = new_x, prob = new_prob), "discrete")
 }
 
+retype_dis_vlaue <- function(f) {
+  x_tbl <- meta_x_tbl(f)
+  # Renormalization of "prob" column will be done inside `new_*()` function
+  new_x_tbl <- data.frame(x = x_tbl[["x"]], prob = x_tbl[["y"]])
+
+  new_pdqr_by_ref(f)(new_x_tbl, type = "discrete")
+}
+
 retype_con <- function(f, method) {
   if (meta_type(f) == "continuous") {
     return(f)
@@ -153,7 +187,8 @@ retype_con <- function(f, method) {
   switch(
     method,
     piecelin = retype_con_piecelin(f),
-    dirac = retype_con_dirac(f)
+    dirac = retype_con_dirac(f),
+    value = retype_con_value(f)
   )
 }
 
@@ -219,4 +254,12 @@ retype_con_dirac <- function(f, h = 1e-8) {
   new_y <- c(   y_zero, x_tbl[["prob"]] / h_vec,    y_zero)
 
   new_pdqr_by_ref(f)(data.frame(x = new_x, y = new_y), "continuous")
+}
+
+retype_con_value <- function(f) {
+  x_tbl <- meta_x_tbl(f)
+  # Renormalization of "y" column will be done inside `new_*()` function
+  new_x_tbl <- data.frame(x = x_tbl[["x"]], y = x_tbl[["prob"]])
+
+  new_pdqr_by_ref(f)(new_x_tbl, type = "continuous")
 }