+++ NEWS.md
@@ -1,5 +1,7 @@
 # pdqr (development version)
 
+* `form_retype()` gets new method "value" that preserves ratio of d-function values at certain "x" points.
+
 # pdqr 0.1.0
 
 * Initial release.